+++ database/migrations/2022_05_04_055047_create_area_table.php
@@ -17,7 +17,7 @@ class CreateAreaTable extends Migration
             $table->id();
             $table->string('name',30)->name('場域名稱');
             $table->string('memo',100)->nullable()->name('備註說明');
-            $table->tinyInteger('status')->name('資料狀態');
+            $table->integer('status')->name('資料狀態');
             $table->timestamps();
             $table->bigInteger('created_by')->name('資料建立人員');
             $table->bigInteger('updated_by')->name('最後編輯人員');
