+++ database/migrations/2022_05_04_055203_create_smartpole_type_table.php
@@ -22,7 +22,7 @@ class CreateSmartpoleTypeTable extends Migration
             $table->float('weight')->nullable()->name('重量(公斤)');
             $table->float('attached_weight')->nullable()->name('可附掛重量(公斤)');
             $table->string('memo',100)->nullable()->name('備註說明');
-            $table->tinyInteger('status')->name('資料狀態');
+            $table->integer('status')->name('資料狀態');
             $table->timestamps();
             $table->bigInteger('created_by')->name('資料建立人員');
             $table->bigInteger('updated_by')->name('最後編輯人員');
