+++ database/migrations/2022_05_04_060814_create_attached_device_table.php
@@ -19,8 +19,8 @@ class CreateAttachedDeviceTable extends Migration
             $table->string('memo',100)->nullable()->name('備註說明');
             $table->integer('status')->name('資料狀態');
             $table->timestamps();
-            $table->integer('created_by')->name('資料建立人員');
-            $table->integer('updated_by')->name('最後編輯人員');
+            $table->bigInteger('created_by')->name('資料建立人員');
+            $table->bigInteger('updated_by')->name('最後編輯人員');
         });
     }
