+++ database/migrations/2022_05_05_004937_create_applyleased_report_remove_file_table.php
@@ -19,7 +19,7 @@ class CreateApplyleasedReportRemoveFileTable extends Migration
             $table->string('file_path',200)->name('存放路徑');
             $table->string('file_name',20)->name('檔案名稱');
             $table->string('memo',100)->nullable()->name('備註說明');
-            $table->bigInteger('status')->name('資料狀態');
+            $table->integer('status')->name('資料狀態');
             $table->timestamps();
             $table->bigInteger('created_by')->name('資料建立人員');
             $table->bigInteger('updated_by')->name('最後編輯人員');
